Move multiview functions to `game/client/multiview.cpp`
