+++ CMakeLists.txt
@@ -2721,6 +2721,7 @@ if(CLIENT)
     laser_data.h
     lineinput.cpp
     lineinput.h
+    multiview.cpp
     pickup_data.cpp
     pickup_data.h
     prediction/entities/character.cpp
